+++ Bgg.Net.Common/RequestHandlers/Guilds/IGuildHandler.cs
@@ -14,7 +14,7 @@ namespace Bgg.Net.Common.RequestHandlers.Guilds
         /// </summary>
         /// <param name="id">The id to retrieve.</param>
         /// <returns>A <see cref="BggResult{T}"/> where T is a <see cref="Guild"/>.</returns>
-        Task<BggResult<Guild>> GetGuildById(int id);
+        Task<BggResult<Guild>> GetGuildById(long id);
 
         /// <summary>
         /// Gets a guild by the given id.  Includes members in the result.
@@ -23,7 +23,7 @@ namespace Bgg.Net.Common.RequestHandlers.Guilds
         /// <param name="sortType">Optional parameter dictate how to sort the members.  Defaults to username.</param>
         /// <param name="page">Optional parameter indicates which page of members to include in the set.  Defaults to 1.</param>
         /// <returns>A <see cref="BggResult{T}"/> where T is a <see cref="Guild"/>.</returns>
-        Task<BggResult<Guild>> GetGuildByIdWithMembers(int id, SortType sortType = SortType.UserName, int page = 1);
+        Task<BggResult<Guild>> GetGuildByIdWithMembers(long id, SortType sortType = SortType.UserName, int page = 1);
 
         /// <summary>
         /// Gets a guild given extensible parameters.
